[Feat] 관심사 목록 조회 구현 및 관심사 수정 개선, Actions workflow 네이버 API 키 추가

+++ src/test/resources/schema.sql
@@ -69,7 +69,7 @@ CREATE TABLE interests
 (
     id               UUID PRIMARY KEY,
     name             VARCHAR(50)              NOT NULL,
-    subscriber_count INTEGER DEFAULT 0        NOT NULL,
+    subscriber_count BIGINT DEFAULT 0        NOT NULL,
     created_at       TIMESTAMP WITH TIME ZONE NOT NULL,
     updated_at       TIMESTAMP WITH TIME ZONE NULL
 );
